Move `StatsTab` to standalone file
