+++ web-app/client/src/components/FilePropsView/FilePropsView.tsx
@@ -1,4 +1,4 @@
-import { FC, useEffect, useReducer, useState } from 'react';
+import { FC, useState } from 'react';
 import { useForm, SubmitHandler, Controller } from 'react-hook-form';
 import _ from 'lodash';
 import Button from '@components/Button';
@@ -7,27 +7,7 @@ import { Tab, TabView } from '@components/TabView/TabView';
 import { FileProps } from 'types/globalTypes';
 import Tooltip from '@components/Tooltip';
 import styles from './FilePropsView.module.scss';
-import { Alert } from '@components/FileStats/Alert';
-import NumberSlider from '@components/Inputs/NumberSlider/NumberSlider';
-import { Progress } from '@components/FileStats/Progress';
-import { Table } from '@components/FileStats/Table';
-import { ColumnCard } from '@components/FileStats/ColumnCard';
-import { useRouter } from 'next/router';
-import { OptionWithBadge } from '@components/FilePropsView/OptionWithBadge';
-import { Menu } from '@components/FilePropsView/Menu';
-import { useMutation, useQuery } from '@apollo/client';
-import { GET_FILE_STATS } from '@graphql/operations/queries/getFileStats';
-import {
-  getFileStats,
-  getFileStatsVariables,
-} from '@graphql/operations/queries/__generated__/getFileStats';
-import { getOverview } from '@utils/fileStats';
-import { START_PROCESSING_STATS } from '@graphql/operations/mutations/startProcessingStats';
-import {
-  startProcessingStats,
-  startProcessingStatsVariables,
-} from '@graphql/operations/mutations/__generated__/startProcessingStats';
-import classNames from 'classnames';
+import { StatsTab } from '@components/FilePropsView/StatsTab';
 
 type Props = {
   data: FileProps;
@@ -176,181 +156,6 @@ const FilePropsForm: FC<Props & FormProps> = ({ data, switchEdit }) => {
         </Button>
       </div>
     </form>
-  );
-};
-
-export type ColumnOption = {
-  value: number;
-  label: string;
-  type?: string;
-  categorical?: boolean;
-};
-
-enum Stage {
-  Start,
-  Processing,
-  Show,
-}
-
-type StatsTabProps = {
-  fileId: string;
-};
-
-const StatsTab: FC<StatsTabProps> = ({ fileId }: StatsTabProps) => {
-  const router = useRouter();
-  const [threadsCount, setThreadsCount] = useState(1);
-  const [stage, setStage] = useState<Stage | null>(null);
-  const [selectedColumn, setSelectedColumn] = useState(-1);
-
-  const {
-    data: fileStats,
-    startPolling,
-    stopPolling,
-  } = useQuery<getFileStats, getFileStatsVariables>(GET_FILE_STATS, {
-    variables: {
-      fileID: fileId,
-    },
-    onCompleted: (fileStats) => {
-      const file = fileStats?.datasetInfo;
-      if (!file) return;
-
-      if (!file.hasStats) return setStage(Stage.Start);
-
-      if (file.statsProgress !== 100) return setStage(Stage.Processing);
-
-      return setStage(Stage.Show);
-    },
-  });
-
-  const [startProcessingStats] = useMutation<
-    startProcessingStats,
-    startProcessingStatsVariables
-  >(START_PROCESSING_STATS, {
-    onCompleted: () => setStage(Stage.Processing),
-  });
-
-  useEffect(() => {
-    if (stage === Stage.Processing) startPolling?.(1000);
-
-    if (stage === Stage.Show) stopPolling?.();
-  }, [stage]);
-
-  const file = fileStats?.datasetInfo;
-
-  if (!file)
-    return (
-      <div className={styles.loading}>
-        <h5>Loading...</h5>
-      </div>
-    );
-
-  const startProcessing = (
-    <>
-      <Alert>
-        Statistics have not been processed yet.
-        <br />
-        Would you like to start processing?
-      </Alert>
-      <NumberSlider
-        sliderProps={{ min: 1, max: 9, step: 1 }}
-        label="Thread count"
-        value={threadsCount}
-        onChange={(value) => setThreadsCount(Math.round(value))}
-      />
-    </>
-  );
-
-  const processing = (
-    <>
-      <div className={styles.processing}>
-        <div className={styles['processing-label']}>
-          <span>Discovering statistics</span>
-          <span>{file.statsProgress}%</span>
-        </div>
-        <Progress value={file.statsProgress} />
-      </div>
-    </>
-  );
-
-  const overview = (
-    <Table>
-      <tbody>
-        {getOverview(file).map((item) => (
-          <tr key={item.name}>
-            <th>{item.name}</th>
-            <td>{item.value}</td>
-          </tr>
-        ))}
-      </tbody>
-    </Table>
-  );
-
-  const options: ColumnOption[] = [
-    { value: -1, label: 'Overview' },
-    ...file.stats.map((column) => ({
-      value: column.columnIndex,
-      label: column.columnName ?? '',
-      type: column.type ?? '',
-      categorical: !!column.isCategorical,
-    })),
-  ];
-
-  const columns = (
-    <>
-      <div className={styles['header-with-select']}>
-        <h5>Columns</h5>
-        <Select
-          isSearchable={false}
-          value={options.find((option) => option.value === selectedColumn)}
-          onChange={(e) => setSelectedColumn((e as ColumnOption).value)}
-          options={options}
-          components={{ Option: OptionWithBadge, Menu }}
-        />
-      </div>
-
-      {selectedColumn === -1 && overview}
-      {selectedColumn !== -1 && (
-        <ColumnCard
-          column={
-            file.stats.find((column) => column.columnIndex === selectedColumn)!
-          }
-          compact
-        />
-      )}
-    </>
-  );
-
-  return (
-    <>
-      <div className={styles.stats}>
-        {stage === Stage.Start && startProcessing}
-        {stage === Stage.Processing && processing}
-        {stage === Stage.Show && columns}
-      </div>
-      <div className={styles.buttonsRow}>
-        {stage !== Stage.Show && (
-          <Button
-            disabled={stage !== Stage.Start}
-            onClick={() =>
-              startProcessingStats({
-                variables: { fileID: fileId, threadsCount },
-              })
-            }
-          >
-            Start Processing
-          </Button>
-        )}
-        {stage === Stage.Show && (
-          <Button
-            onClick={() =>
-              router.push(`/create-task/file-stats?fileId=${fileId}`)
-            }
-          >
-            Show More
-          </Button>
-        )}
-      </div>
-    </>
   );
 };
